+++ test/expected/def-file-var.js
@@ -1,0 +1,6 @@
+// fixtures/def-file-var.js
+// _fixtures/def-file-var.js $__FILE $FILE _FILE$
+export default function main () {
+  // out from "fixtures/def-file-var.js"
+}
+//fixtures/def-file-var.js