+++ test/fixtures/def-file-var.js
@@ -1,0 +1,6 @@
+// $_FILE
+// _$_FILE $__FILE $FILE _FILE$
+export default function main () {
+  // out from "$_FILE"
+}
+//$_FILE