+++ test/fixtures/file-var.js
@@ -1,7 +1,0 @@
-
-// fixtures/file-var.js
-// _fixtures/file-var.js $__FILE $FILE _FILE$
-export default function main () {
-  // out from "fixtures/file-var.js"
-}
-//fixtures/file-var.js